// src/app/shared/enums/routes.enum.ts
export enum AppRoutes {
  NotFound = '404',
  Other = '**',
  Empty = '',
  Home = 'home',
  Contact = 'contact-us',
}

export enum cartRoutes {
  ShoppingCart = 'shopping-cart',
  Checkout = 'checkout',
}

export enum productRoutes {
  ProductList = 'product-list',
  ProductOverView = 'product-overview',
}